Move apple alias tokens out
