+++ packages/theming/apple-theme/package.json
@@ -34,7 +34,6 @@
     "@fluentui-react-native/default-theme": ">=0.9.13 <1.0.0",
     "@fluentui-react-native/memo-cache": "^1.1.7",
     "@fluentui-react-native/theme": ">=0.6.6 <1.0.0",
-    "@fluentui-react-native/theme-tokens": "^0.13.2",
     "@fluentui-react-native/theme-types": ">=0.14.1 <1.0.0",
     "@fluentui-react-native/theming-utils": "^0.9.2"
   },
