+++ packages/theming/apple-theme/src/createMacOSAliasTokens.ts
@@ -1,4 +1,4 @@
-import { getMacOSAliasTokens } from '@fluentui-react-native/theme-tokens';
+import { getMacOSAliasTokens } from './getMacOSAliasTokens';
 import { AliasColorTokens, AppearanceOptions } from '@fluentui-react-native/theme-types';
 import { mapPipelineToTheme } from '@fluentui-react-native/theming-utils';
 import { memoize } from '@fluentui-react-native/memo-cache';
