+++ packages/theming/apple-theme/src/dark-high-contrast-macos/tokens-alias.ts
@@ -1,0 +1,3 @@
+import aliasTokens from '../generated/dark-high-contrast-macos/tokens-aliases.json';
+
+export default aliasTokens;
