+++ packages/theming/apple-theme/src/dark-macos/tokens-alias.ts
@@ -1,0 +1,3 @@
+import aliasTokens from '../generated/dark-macos/tokens-aliases.json';
+
+export default aliasTokens;
