+++ packages/theming/apple-theme/src/light-macos/tokens-alias.ts
@@ -1,0 +1,3 @@
+import aliasTokens from '../generated/light-macos/tokens-aliases.json';
+
+export default aliasTokens;
