--- packages/theming/theme-tokens/src/dark-high-contrast-macos/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '../generated/dark-high-contrast-macos/reactnative/tokens-aliases.json';
-
-export default aliasTokens;
