--- packages/theming/theme-tokens/src/dark-macos/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '../generated/dark-macos/reactnative/tokens-aliases.json';
-
-export default aliasTokens;
