+++ packages/theming/theme-tokens/src/index.ts
@@ -1,3 +1,2 @@
 export { default as globalTokens } from './tokens-global';
 export { getAliasTokens } from './getAliasTokens';
-export { getMacOSAliasTokens } from './getMacOSAliasTokens';
