--- packages/theming/theme-tokens/src/light-high-contrast-macos/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '../generated/light-high-contrast-macos/reactnative/tokens-aliases.json';
-
-export default aliasTokens;
