--- packages/theming/theme-tokens/src/light-macos/tokens-alias.ts
@@ -1,3 +1,0 @@
-import aliasTokens from '../generated/light-macos/reactnative/tokens-aliases.json';
-
-export default aliasTokens;
